+++ src/main/java/com/student/book_advisor/dto/auxiliaryDTOs/MyBooksReadDTO.java
@@ -1,19 +1,19 @@
 package com.student.book_advisor.dto.auxiliaryDTOs;
 
 public class MyBooksReadDTO {
-    private Long myBooksID;
+    private Integer myBooksID;
     private String title;
 
-    public MyBooksReadDTO(Long myBooksID, String title) {
+    public MyBooksReadDTO(Integer myBooksID, String title) {
         this.myBooksID = myBooksID;
         this.title = title;
     }
 
-    public Long getMyBooksID() {
+    public Integer getMyBooksID() {
         return myBooksID;
     }
 
-    public void setMyBooksID(Long myBooksID) {
+    public void setMyBooksID(Integer myBooksID) {
         this.myBooksID = myBooksID;
     }
