+++ src/main/java/com/student/book_advisor/entityRepositories/SagaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SagaRepository extends JpaRepository<Saga,Long> {
+public interface SagaRepository extends JpaRepository<Saga,Integer> {
 
     public Saga findByBook(Libro book);
 }
